v1.1.1 - Modernization/Filter Dictionary
- Updated tooling to match other 1.1 clients
- Added filter dictionary endpoint method
