--- tslint.json
@@ -1,9 +1,0 @@
-{
-    "extends": "tslint-config-airbnb",
-    "rules": {
-        "max-line-length": {
-            "options": [300]
-        },
-        "object-literal-shorthand": [true, "never"]
-    }
-}
